requirements.txt
psycopg2~=2.8.6
python-dotenv~=0.15.0
requests~=2.24.0
beautifulsoup4~=4.9.1
deepdiff~=5.0.2
argparse~=1.4.0
python-telegram-bot
html5lib